requirements.txt
PySide6>=6.6,<7.0
PySide6-WebEngine>=6.6,<7.0
Pillow>=10.0
python-pptx>=0.6.21
python-docx>=1.1.0
lxml>=4.9.0
